requirements.txt
streamlit
openai
requests
python-dotenv
tavily-python
firecrawl-py